OpAmp Package Update Glue (#567)
Switched PackageStatuses yaml to a JSON file to prevent partial reads by Updater.
Removed excess fields for package status. We should be able to communicate with available status and error message.
If just started an install, will prevent another PackagesAvailable message from starting another install.
If OpAMP client errors out at any point, sets the status to failed with an error message (if possible) in the JSON file.
This will allow the updater to quickly shut down the collector and start up the rollback one (which will then send the message to BindPlane).
On BindPlane connect, will check if the status is installing. If so, will check if Server version matches current version. Based on this will either set status to success or fail and write to JSON file for BindPlane to notice. It should only try to send a message immediately to BindPlane if it was a success.